feat: add split factor for all to all

+++ flink-runtime/src/main/java/org/apache/flink/runtime/executiongraph/IntermediateResult.java
@@ -354,7 +354,7 @@ public class IntermediateResult {
         }
     }
 
-    public int getSplitFactor() {
+    public double getSplitFactor() {
         return intermediateDataSet.getSplitFactor();
     }
