+++ flink-runtime/src/main/java/org/apache/flink/runtime/jobgraph/IntermediateDataSet.java
@@ -96,7 +96,7 @@ public class IntermediateDataSet implements java.io.Serializable {
         return resultType;
     }
 
-    public int getSplitFactor() {
+    public double getSplitFactor() {
         return producer.getSplitFactor();
     }
